created selector folder and moved mood selectors into a file called moodSelectors.js
